BugFix sulla stampa della dimensione del body di un HttpData.

+++ src/main/java/it/uniroma2/sii/util/data/http/HttpData.java
@@ -195,10 +195,10 @@ public abstract class HttpData extends Data {
 	}
 
 	public String toString() {
-		double bodyLength = messageBody.length / 1024;
+		double bodyLength = (double) messageBody.length/(double) 1024;
 		return String.format("%s%sBody Dimension: %sKB\n", startLine.toString(),
 				headers.toString(),
 				messageBody != null ? String.valueOf(bodyLength)
-						: String.valueOf(0));
+						: String.valueOf(0.0));
 	}
 }
